+++ frost-core/CHANGELOG.md
@@ -7,6 +7,8 @@ Entries are listed in reverse chronological order.
 
 * Added a new ciphersuite `frost-secp256k1-evm` that uses the secp256k1
   elliptic curve with hash keccak256.
+* Added no-std support for frost-ed448 crate. This became possible after migration to `ed448-goldilocks-plus` (fork of
+  `ed448-goldilocks`).
 
 ## 3.0.0